+++ bootstrap/providers.php
@@ -1,9 +1,11 @@
 <?php
 
 use App\Providers\AppServiceProvider;
+use App\Providers\AuthServiceProvider;
 use App\Providers\VoltServiceProvider;
 
 return [
     AppServiceProvider::class,
+    AuthServiceProvider::class,
     VoltServiceProvider::class,
 ];